+++ demo/desktop-starter/CHANGELOG.md
@@ -1,5 +1,12 @@
 # mandu-desktop-starter
 
+## 0.1.60
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @mandujs/core@0.54.10
+
 ## 0.1.59
 
 ### Patch Changes